# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "cse341 project",
    "description": "BYUI Course API",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/post": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "Post Title"
                },
                "slug": {
                  "example": "this-is-a-post-slug"
                },
                "published": {
                  "example": true
                },
                "tags": {
                  "example": ["a", "b"]
                },
                "author": {
                  "example": "wmezadev"
                },
                "excerpt": {
                  "example": "this is my second post"
                },
                "content": {
                  "example": "Hello World this could be rich text"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "wilmeza"
                },
                "password": {
                  "example": "password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Auth error, email or password not valid"
          }
        }
      }
    },
...
  }
}